+++ common/src/sdk/models/operations/index.ts
@@ -3,9 +3,11 @@
  */
 
 export * from "./configuresupplementaldata";
+export * from "./createapikey";
 export * from "./createcompany";
 export * from "./createdataconnection";
 export * from "./createrule";
+export * from "./deleteapikey";
 export * from "./deletecompany";
 export * from "./deletecompanyconnection";
 export * from "./getcompany";
@@ -21,6 +23,7 @@ export * from "./getpulloperation";
 export * from "./getpushoperation";
 export * from "./getsupplementaldataconfiguration";
 export * from "./getwebhook";
+export * from "./listapikeys";
 export * from "./listcompanies";
 export * from "./listcompanyconnections";
 export * from "./listintegrations";