+++ common/RELEASES.md
@@ -641,3 +641,13 @@ Based on:
 - [typescript v0.36.0] common
 ### Releases
 - [NPM v0.36.0] https://www.npmjs.com/package/@codat/common/v/0.36.0 - common
+
+## 2023-08-23 00:03:32
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.74.7 (2.86.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.36.1] common
+### Releases
+- [NPM v0.36.1] https://www.npmjs.com/package/@codat/common/v/0.36.1 - common